refactor: 기존의 types 구조 .d.ts 방식으로 통합 및 폴더 정리
