+++ src/types/index.d.ts
@@ -1,0 +1,12 @@
+// API 관련 타입들
+export * from './api/leaderBoard';
+export * from './api/user';
+export * from './api/subscription';
+export * from './api/usageLog';
+export * from './api/paymentMethod';
+export * from './api/apiResponse';
+
+// 글로벌 타입들
+export * from './statistics';
+export * from './cycle';
+export * from './error';
